+++ packages/falcon-server/src/index.js
@@ -9,13 +9,16 @@ const { ApolloServer } = require('apollo-server-koa');
 const { EventEmitter2 } = require('eventemitter2');
 const GraphQLJSON = require('graphql-type-json');
 const cors = require('@koa/cors');
+const Cookies = require('cookies');
 const Koa = require('koa');
 const Router = require('koa-router');
 const session = require('koa-session');
+const SessionContext = require('koa-session/lib/context');
 const body = require('koa-body');
 const get = require('lodash/get');
 const capitalize = require('lodash/capitalize');
 const ApiContainer = require('./containers/ApiContainer');
+const ComponentContainer = require('./containers/ComponentContainer');
 const ExtensionContainer = require('./containers/ExtensionContainer');
 const EndpointContainer = require('./containers/EndpointContainer');
 const DynamicRouteResolver = require('./resolvers/DynamicRouteResolver');
@@ -30,8 +33,10 @@ class FalconServer {
     this.loggableErrorCodes = [codes.INTERNAL_SERVER_ERROR, codes.GRAPHQL_PARSE_FAILED];
     this.config = config;
     this.server = null;
+    this.httpServer = null;
     this.cache = null;
     this.backendConfig = {};
+    this.components = {};
     const { maxListeners = 20, verboseEvents = false } = this.config;
     if (config.logLevel) {
       Logger.setLogLevel(config.logLevel);
@@ -60,6 +65,8 @@ class FalconServer {
 
   async initialize() {
     await this.eventEmitter.emitAsync(Events.BEFORE_INITIALIZED, this);
+    this.cache = new Cache(this.getCacheProvider());
+    await this.initializeComponents();
     await this.initializeServerApp();
     await this.initializeContainers();
     await this.initializeApolloServer();
@@ -68,32 +75,63 @@ class FalconServer {
   }
 
   async getApolloServerConfig() {
-    this.cache = new Cache(this.getCacheProvider());
-    const dynamicRouteResolver = new DynamicRouteResolver(this.extensionContainer);
+    this.apolloServerConfig = await this.extensionContainer.createGraphQLConfig(this.getInitialGraphQLConfig());
 
-    const apolloServerConfig = await this.extensionContainer.createGraphQLConfig({
+    /* eslint-disable no-underscore-dangle */
+    // Removing "placeholder" (_) fields from the Type definitions
+    delete this.apolloServerConfig.schema._subscriptionType._fields._;
+
+    // If there were no other fields defined for Type by any other extension
+    // - we need to remove it completely in order to comply with GraphQL specification
+    if (!Object.keys(this.apolloServerConfig.schema._subscriptionType._fields).length) {
+      this.apolloServerConfig.schema._subscriptionType = undefined;
+      delete this.apolloServerConfig.schema._typeMap.Subscription;
+    }
+    /* eslint-enable no-underscore-dangle */
+
+    return this.apolloServerConfig;
+  }
+
+  getInitialGraphQLConfig() {
+    return {
       schemas: [BaseSchema],
       dataSources: () => {
         this.logger.debug('Instantiating GraphQL DataSources');
         const dataSources = {};
         this.apiContainer.dataSources.forEach((value, key) => {
-          dataSources[key] = value(apolloServerConfig);
+          dataSources[key] = value(this.apolloServerConfig);
         });
         return dataSources;
       },
-      schemaDirectives,
+      schemaDirectives: this.getDefaultSchemaDirectives(),
       formatError: error => this.formatGraphqlError(error),
       // inject session and headers into GraphQL context
-      context: ({ ctx }) => ({
-        cache: this.cache,
-        config: this.config,
-        headers: ctx.req.headers,
-        session: ctx.req.session
-      }),
+      context: ({ ctx, connection }) => {
+        const context = {
+          cache: this.cache,
+          config: this.config,
+          components: this.componentContainer.components
+        };
+
+        // Subscription request
+        if (connection) {
+          return {
+            ...context,
+            ...connection.context
+          };
+        }
+
+        // Query/Mutation request
+        return {
+          ...context,
+          headers: ctx.req.headers,
+          session: ctx.req.session
+        };
+      },
       cache: this.cache,
       resolvers: {
         Query: {
-          url: async (...params) => dynamicRouteResolver.fetchUrl(...params),
+          url: async (...params) => this.dynamicRouteResolver.fetchUrl(...params),
           backendConfig: async (...params) => this.fetchBackendConfig(...params)
         },
         Mutation: {
@@ -104,27 +142,18 @@ class FalconServer {
         },
         JSON: GraphQLJSON
       },
+      subscriptions: this.getSubscriptionsOptions(),
       tracing: this.config.debug,
       playground: this.config.debug && {
         settings: {
           'request.credentials': 'include' // include to keep the session between requests
         }
       }
-    });
+    };
+  }
 
-    /* eslint-disable no-underscore-dangle */
-    // Removing "placeholder" (_) fields from the Type definitions
-    delete apolloServerConfig.schema._subscriptionType._fields._;
-
-    // If there were no other fields defined for Type by any other extension
-    // - we need to remove it completely in order to comply with GraphQL specification
-    if (!Object.keys(apolloServerConfig.schema._subscriptionType._fields).length) {
-      apolloServerConfig.schema._subscriptionType = undefined;
-      delete apolloServerConfig.schema._typeMap.Subscription;
-    }
-    /* eslint-enable no-underscore-dangle */
-
-    return apolloServerConfig;
+  getDefaultSchemaDirectives() {
+    return schemaDirectives;
   }
 
   /**
@@ -138,6 +167,7 @@ class FalconServer {
 
     this.router = new Router();
 
+    this.app.context.components = this.componentContainer.components;
     this.app.use(body());
     this.app.use(
       cors({
@@ -180,6 +210,16 @@ class FalconServer {
 
     this.endpointContainer = new EndpointContainer(this.eventEmitter);
     await this.endpointContainer.registerEndpoints(this.config.endpoints);
+
+    this.dynamicRouteResolver = new DynamicRouteResolver(this.extensionContainer);
+  }
+
+  async initializeComponents() {
+    await this.eventEmitter.emitAsync(Events.BEFORE_COMPONENT_CONTAINER_CREATED, this.config.components);
+    /** @type {ComponentContainer} */
+    this.componentContainer = new ComponentContainer(this.eventEmitter);
+    await this.componentContainer.registerComponents(this.config.components);
+    await this.eventEmitter.emitAsync(Events.AFTER_COMPONENT_CONTAINER_CREATED, this.componentContainer);
   }
 
   /**
@@ -280,6 +320,10 @@ class FalconServer {
     };
   }
 
+  isSubscriptionsServerRequired() {
+    return 'Subscription' in this.apolloServerConfig.schema.getTypeMap();
+  }
+
   start() {
     const handleStartupError = err => {
       this.eventEmitter.emitAsync(Events.ERROR, err).then(() => {
@@ -295,17 +339,78 @@ class FalconServer {
       .then(
         () =>
           new Promise(resolve => {
-            this.app.listen({ port: this.config.port }, () => {
+            this.httpServer = this.app.listen({ port: this.config.port }, () => {
               this.logger.info(`🚀 Server ready at http://localhost:${this.config.port}`);
               this.logger.info(
                 `🌍 GraphQL endpoint ready at http://localhost:${this.config.port}${this.server.graphqlPath}`
               );
+              this.startSubscriptionsServer();
               resolve();
             });
           }, handleStartupError)
       )
       .then(() => this.eventEmitter.emitAsync(Events.AFTER_STARTED, this))
       .catch(handleStartupError);
+  }
+
+  /**
+   * Starts (if needed) the subscriptions server (based on the stitched GraphQL Schema - Subscription type has resolvers)
+   */
+  startSubscriptionsServer() {
+    if (this.isSubscriptionsServerRequired()) {
+      this.server.installSubscriptionHandlers(this.httpServer);
+      this.logger.info(
+        `🔌 GraphQL Subscriptions endpoint ready at ws://localhost:${this.config.port}${this.server.subscriptionsPath}`
+      );
+    }
+  }
+
+  /**
+   * Get default Subscriptions Options with even handlers to properly initialize required context values
+   */
+  getSubscriptionsOptions() {
+    return {
+      onConnect: (connectionParams, websocket, context) => {
+        const { cookie: paramsCookie } = connectionParams;
+        let { request } = context;
+
+        if (paramsCookie) {
+          // Faking `cookie` data passed from `connectionParams` to the request "headers" object
+          request = {
+            headers: {
+              cookie: paramsCookie,
+              ...context.request.headers
+            }
+          };
+        }
+
+        // Checking signed cookies (without a `res` argument, since for Subscriptions there're no traditional "responses")
+        try {
+          const cookies = new Cookies(
+            request,
+            {},
+            {
+              keys: this.config.session.keys
+            }
+          );
+          // Manually decrypting session cookie
+          const sessionContext = new SessionContext(
+            {
+              sessionOptions: {},
+              cookies
+            },
+            this.config.session.options
+          );
+
+          return {
+            headers: context.request.headers,
+            session: sessionContext.get()
+          };
+        } catch (e) {
+          throw new Error('Failed to parse Cookie');
+        }
+      }
+    };
   }
 }